bottom bar updated
