+++ lib/presentation/navigation/app_router.gr.dart
@@ -8,14 +8,34 @@
 // coverage:ignore-file
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:auto_route/auto_route.dart' as _i3;
-import 'package:canteen/presentation/pages/auth/login_page.dart' as _i1;
-import 'package:canteen/presentation/pages/auth/signup_page.dart' as _i2;
+import 'package:auto_route/auto_route.dart' as _i4;
+import 'package:canteen/presentation/pages/auth/login_page.dart' as _i2;
+import 'package:canteen/presentation/pages/auth/signup_page.dart' as _i3;
+import 'package:canteen/presentation/pages/home/home_page.dart' as _i1;
 
 /// generated route for
-/// [_i1.LoginPage]
-class LoginRoute extends _i3.PageRouteInfo<void> {
-  const LoginRoute({List<_i3.PageRouteInfo>? children})
+/// [_i1.HomePage]
+class HomeRoute extends _i4.PageRouteInfo<void> {
+  const HomeRoute({List<_i4.PageRouteInfo>? children})
+      : super(
+          HomeRoute.name,
+          initialChildren: children,
+        );
+
+  static const String name = 'HomeRoute';
+
+  static _i4.PageInfo page = _i4.PageInfo(
+    name,
+    builder: (data) {
+      return const _i1.HomePage();
+    },
+  );
+}
+
+/// generated route for
+/// [_i2.LoginPage]
+class LoginRoute extends _i4.PageRouteInfo<void> {
+  const LoginRoute({List<_i4.PageRouteInfo>? children})
       : super(
           LoginRoute.name,
           initialChildren: children,
@@ -23,18 +43,18 @@ class LoginRoute extends _i3.PageRouteInfo<void> {
 
   static const String name = 'LoginRoute';
 
-  static _i3.PageInfo page = _i3.PageInfo(
+  static _i4.PageInfo page = _i4.PageInfo(
     name,
     builder: (data) {
-      return const _i1.LoginPage();
+      return const _i2.LoginPage();
     },
   );
 }
 
 /// generated route for
-/// [_i2.SignUpPage]
-class SignUpRoute extends _i3.PageRouteInfo<void> {
-  const SignUpRoute({List<_i3.PageRouteInfo>? children})
+/// [_i3.SignUpPage]
+class SignUpRoute extends _i4.PageRouteInfo<void> {
+  const SignUpRoute({List<_i4.PageRouteInfo>? children})
       : super(
           SignUpRoute.name,
           initialChildren: children,
@@ -42,10 +62,10 @@ class SignUpRoute extends _i3.PageRouteInfo<void> {
 
   static const String name = 'SignUpRoute';
 
-  static _i3.PageInfo page = _i3.PageInfo(
+  static _i4.PageInfo page = _i4.PageInfo(
     name,
     builder: (data) {
-      return const _i2.SignUpPage();
+      return const _i3.SignUpPage();
     },
   );
 }
