+++ lib/presentation/pages/auth/signup_page.dart
@@ -2,10 +2,10 @@ import 'package:auto_route/annotations.dart';
 import 'package:auto_route/auto_route.dart';
 import 'package:canteen/core/validators.dart';
 import 'package:canteen/presentation/navigation/app_router.dart';
-import 'package:canteen/presentation/widgets/app_button.dart';
+import 'package:canteen/presentation/widgets/buttons/app_button.dart';
 import 'package:flutter/material.dart';
-import 'package:canteen/presentation/widgets/role_dropdown.dart';
-import 'package:canteen/presentation/widgets/app_text_form_field.dart';
+import 'package:canteen/presentation/widgets/fields/role_dropdown.dart';
+import 'package:canteen/presentation/widgets/fields/app_text_form_field.dart';
 
 @RoutePage()
 class SignUpPage extends StatefulWidget {
@@ -84,6 +84,7 @@ class _SignUpPageState extends State<SignUpPage> {
                       const SnackBar(content: Text('Please select a role')),
                     );
                   } else {
+                    context.router.pushNamed('/home');
                   }
                 },
               ),
