+++ lib/presentation/pages/home/debts/debts_page.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class DebtsPage extends StatelessWidget {
+  const DebtsPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text(
+        "Welcome to the Debts Page!",
+        style: TextStyle(
+          fontSize: 18,
+          color: Colors.black,
+        ),
+      ),
+    );
+  }
+}
