+++ lib/presentation/pages/home/delivery/delivery_page.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class DeliveryPage extends StatelessWidget {
+  const DeliveryPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text(
+        "Welcome to the Delivery Page!",
+        style: TextStyle(
+          fontSize: 18,
+          color: Colors.black,
+        ),
+      ),
+    );
+  }
+}
