+++ lib/presentation/pages/home/recomendations/recommendations_page.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class RecommendationsPage extends StatelessWidget {
+  const RecommendationsPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text(
+        "Welcome to the Recommendations Page!",
+        style: TextStyle(
+          fontSize: 18,
+          color: Colors.black,
+        ),
+      ),
+    );
+  }
+}
