+++ lib/presentation/widgets/layout/navigation_bar.dart
