+++ pubspec.yaml
@@ -21,6 +21,7 @@ dependencies:
   auto_route_generator: ^9.0.0
   build_runner: ^2.4.13
   get_it: ^8.0.2
+  curved_navigation_bar: ^1.0.6
 
 dev_dependencies:
   flutter_test:
@@ -33,6 +34,7 @@ flutter:
   assets:
     - assets/images/
     - assets/icons/
+    - assets/layouts/
 
   # fonts:
   #   - family: Schyler
